Move `Channel` to lib.rs
