+++ src/lib.rs
@@ -8,9 +8,10 @@ mod transaction;
 pub mod update;
 
 use crate::{
-    keys::{OwnershipPublicKey, PublishingKeyPair, RevocationKeyPair},
-    transaction::{CommitTransaction, SplitTransaction},
+    keys::{OwnershipKeyPair, OwnershipPublicKey, PublishingKeyPair, RevocationKeyPair},
+    transaction::{CommitTransaction, FundingTransaction, SplitTransaction},
 };
+use anyhow::bail;
 use bitcoin::Amount;
 use ecdsa_fun::adaptor::EncryptedSignature;
 use keys::{PublishingPublicKey, RevocationPublicKey, RevocationSecretKey};
@@ -48,4 +49,59 @@ pub struct RevokedState {
 pub struct ChannelBalance {
     a: (Amount, OwnershipPublicKey),
     b: (Amount, OwnershipPublicKey),
+}
+
+pub struct Channel {
+    x_self: OwnershipKeyPair,
+    X_other: OwnershipPublicKey,
+    TX_f_body: FundingTransaction,
+    current_state: ChannelState,
+    revoked_states: Vec<RevokedState>,
+}
+
+#[derive(Clone, Copy, Debug, PartialEq)]
+pub struct LocalBalance {
+    pub ours: Amount,
+    pub theirs: Amount,
+}
+
+impl Channel {
+    pub fn new(party: create::Party6) -> Self {
+        Self {
+            x_self: party.x_self,
+            X_other: party.X_other,
+            TX_f_body: party.TX_f_body,
+            current_state: ChannelState {
+                TX_c: party.TX_c,
+                encsig_TX_c_self: party.encsig_TX_c_self,
+                encsig_TX_c_other: party.encsig_TX_c_other,
+                r_self: party.r_self,
+                R_other: party.R_other,
+                y_self: party.y_self,
+                Y_other: party.Y_other,
+                signed_TX_s: party.signed_TX_s,
+            },
+            revoked_states: Vec::new(),
+        }
+    }
+
+    pub fn balance(&self) -> anyhow::Result<LocalBalance> {
+        let channel_balance = self.current_state.signed_TX_s.balance();
+
+        match channel_balance {
+            ChannelBalance {
+                a: (ours, X_a),
+                b: (theirs, X_b),
+            } if X_a == self.x_self.public() && X_b == self.X_other => {
+                Ok(LocalBalance { ours, theirs })
+            }
+            ChannelBalance {
+                a: (theirs, X_a),
+                b: (ours, X_b),
+            } if X_a == self.X_other && X_b == self.x_self.public() => {
+                Ok(LocalBalance { ours, theirs })
+            }
+            _ => bail!("split transaction does not pay to X_self and X_other"),
+        }
+    }
 }
